Episode 1 of The Book of Boba Fett commentary.
