+++ posts/2020/2020-10-31-The-Mandalorian-Season-One.md
@@ -1,7 +1,7 @@
 ---
 title: "The Mandalorian: Season One - A Few Thoughts"
 category: Exactly One Hobby
-series: The Mandalorian
+series: Star Wars Commentary
 tags:
 - starwars
 - commentary
@@ -11,7 +11,7 @@ date: 2020-10-31
 
 I watched season one of _The Mandalorian_ as it was released a year ago. I had **_a lot_** of thoughts on the matter at the time, so going into this second season, I wanted to go back and rewatch the first season. This time, writing down all the highlights that I could remember from the first go around. You'll find that I was fond of this show. There were some entertaining pieces, but overall I found it to basically just be an episodic gunslinger-in-space derivative. There's nothing particularly unique or extraordinary about the show. It's half-decent entertainment, so long as you ignore continuity errors, dry dialog, and lots of tropes. There are spoilers throughout, but I don't care.
 
-> The first in a series of posts on ['My Thoughts on _The Mandalorian_'](/the-mandalorian-series). A new thread after each season!
+> The first in a series of posts on ['My Thoughts on _The Mandalorian_'](/star-wars-commentary) and other _Star Wars_ series. A new thread after each season!
 
 ## Episode 1: The Gunslinger... er... Mandalorian
