+++ posts/2021/2021-12-29-The-Book-of-Boba-Fett-Season-1.md
@@ -1,0 +1,30 @@
+---
+title: "The Book of Boba Fett: Season One - A Few Thoughts"
+category: Exactly One Hobby
+series: Star Wars Commentary
+tags:
+- starwars
+- commentary
+- disneyplus
+date: 2021-12-29
+---
+
+As I watched the first two seasons of _The Mandalorian_, I had **_a lot_** of thoughts on the matter. Mostly that the show was derivative, loaded with tropes and cliches, monotonous in pacing, but with a stellar score. As _The Book of Boba Fett_ was being pitched, I had a feeling that I would also have Thoughts on the Matter. In the immortal words of Rick O'Connell, "here we go again."
+
+![Here we go again.](/images/mummy3-here-we-go-again.gif)
+
+> The first in a series of posts on ['My Thoughts on _TBoBF_'](/star-wars-commentary) and other _Star Wars_ series. A new thread after each season!
+
+## Episode 1: Stranger in a Strange Series
+
+- Talk about a bad case of heartburn.
+- Wow... such original cinematogray, a hand shooting out of the ground! First off, the impracticality of being able to shove your arm past the elbow into the air without having disturbed any other part of the ground, especially with a big-ass helmet covering your dumb noggin'.
+- I always thought the tiny t-shaped visor slot was just for show - that they actually had broad visibility in the helmet. But that shot tells me the Mandalorians really have shitty visibility.
+- How can he have such perfectly white teeth having been malnourished, dehydrated, and nearly dead for... however long it's been. Even if this was just one day - there's no way he has a porcelain grill.
+- If you're the crime boss, why do you think _anyone_ would keep their word; especially old bodyguards.
+- I can't even describe how uninteresting this fight scene is... how do they manage to make it so boring?
+- Space parkour ninjas! They're all the rage.
+- I get pretty pissed off, too, when someone disturbs my slumber.
+- A simple act of kindness always sparks another.
+
+At first, I thought this episode was going to be tolerable; the longer it went on, it dipped into terrible. By the time the episode was over... well, [Drew Magary](https://www.sfgate.com/culture-columns/article/The-Book-Of-Boba-Fett-Premiere-Was-The-Worst-16738851.php) was much more eloquent about what an abysmal failure this episode was.
